wip invaders part 3

+++ invaders/src/render.rs
@@ -9,8 +9,8 @@ use crossterm::{
 
 use crate::frame::Frame;
 
-fn queue_write_to(stdout: &mut Stdout, x: u16, y: u16, s: &str) {
-    stdout.queue(MoveTo(x, y)).unwrap();
+fn queue_write_to(stdout: &mut Stdout, x: usize, y: usize, s: &str) {
+    stdout.queue(MoveTo(x as u16, y as u16)).unwrap();
     print!("{}", s);
 }
 
@@ -21,7 +21,7 @@ pub fn clear_and_render(stdout: &mut Stdout, curr_frame: &Frame) {
 
     for (x, col) in curr_frame.iter().enumerate() {
         for (y, &s) in col.iter().enumerate() {
-            queue_write_to(stdout, x as u16, y as u16, s);
+            queue_write_to(stdout, x, y, s);
         }
     }
     stdout.flush().unwrap();
@@ -31,7 +31,7 @@ pub fn render(stdout: &mut Stdout, last_frame: &Frame, curr_frame: &Frame) {
     for (x, col) in curr_frame.iter().enumerate() {
         for (y, &s) in col.iter().enumerate() {
             if s != last_frame[x][y] {
-                queue_write_to(stdout, x as u16, y as u16, s);
+                queue_write_to(stdout, x, y, s);
             }
         }
     }
